refactor(providers): split Nostr Apps providers out of app_providers.dart
Second batch of #4506. Moves the five Nostr Apps providers and their
private `_AuthServiceBridgeAdapter` to a new
`lib/providers/nostr_apps_providers.dart`:

- nostrAppDirectoryServiceProvider
- nostrAppGrantStoreProvider
- nostrAppBridgePolicyProvider
- nostrAppAuditServiceProvider
- nostrAppBridgeServiceProvider

`app_providers.dart` adds `export 'nostr_apps_providers.dart';` so the
seven consumer files (4 lib + 3 test) keep their existing
`import 'package:openvine/providers/app_providers.dart'` imports.

The new file imports `app_providers.dart` to reach `authServiceProvider`
and `nip98AuthServiceProvider`, which still live there until batches 7
and 5 respectively. Removed now-unused
`package:nostr_app_bridge_repository/nostr_app_bridge_repository.dart`
and `package:openvine/config/app_config.dart` imports from
`app_providers.dart`.

No `@Riverpod` annotations in this batch — codegen output unchanged.

Refs #4506, #4339 Wave 2
